Add limitation to UPDATE/DELETE methods for users who did not create these resources

+++ RestaurantAPI/Services/RestaurantService.cs
@@ -1,8 +1,11 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
+using RestaurantAPI.Authorization;
 using RestaurantAPI.Entities;
 using RestaurantAPI.Exceptions;
 using RestaurantAPI.Models;
+using System.Security.Claims;
 
 namespace RestaurantAPI.Services
 {
@@ -11,22 +14,35 @@ namespace RestaurantAPI.Services
         private readonly RestaurantDbContext _dbContext;
         private readonly IMapper _mapper;
         private readonly ILogger<RestaurantService> _logger;
+        private readonly IAuthorizationService _auhtorizationService;
+        private readonly IUserContextService _userContextService;
 
-        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger)
+        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger
+            , IAuthorizationService authorizationService, IUserContextService userContextService)
         {
             _dbContext = dbContext;
             _mapper = mapper;
             _logger = logger;
+            _auhtorizationService = authorizationService;
+            _userContextService = userContextService;
         }
 
         public async Task UpdateAsync(int id, UpdateRestaurantDto dto)
         {
+
             var restaurant = await _dbContext
               .Restaurants
               .FirstOrDefaultAsync(r => r.Id == id);
 
             if (restaurant is null)
                 throw new NotFoundException("Restaurant not found");
+
+            var authorizationResult = await _auhtorizationService.AuthorizeAsync(_userContextService.User, restaurant, new ResourceOperationRequirement(ResourceOperation.Update));
+
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException("Update of this resource is forbidden");
+            }
 
             restaurant.Name = dto.Name;
             restaurant.Description = dto.Description;
@@ -46,6 +62,13 @@ namespace RestaurantAPI.Services
 
             if(restaurant is null)
                 throw new NotFoundException("Restaurant not found");
+
+            var authorizationResult = await _auhtorizationService.AuthorizeAsync(_userContextService.User, restaurant, new ResourceOperationRequirement(ResourceOperation.Delete));
+
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException("Delete of this resource is forbidden");
+            }
 
             _dbContext.Restaurants.Remove(restaurant);
             await _dbContext.SaveChangesAsync();
@@ -94,6 +117,7 @@ namespace RestaurantAPI.Services
         public async Task<int> CreateAsync(CreateRestaurantDto dto)
         {
             var restaurant = _mapper.Map<Restaurant>(dto);
+            restaurant.CreatedById = _userContextService.GetUserId;
             await _dbContext.Restaurants.AddAsync(restaurant);
             await _dbContext.SaveChangesAsync();
